test cron output echo
