+++ .github/workflows/cron-job.yml
@@ -31,33 +31,27 @@ jobs:
         id: check-update
         continue-on-error: true
         run: |
-          for file in ${{ steps.fetch.outputs.paths }}; do
-            if [[ $file == *".json"* ]]; then
-                yarn ts-node src/index.ts $file
-                echo "${{ steps.check-update.outputs.errorOutput }}"
-            fi
-          done
+          echo yarn ts-node src/schedule.ts
 
-      - name: Slack Alert Notification
-        uses: slackapi/slack-github-action@v1.23.0
-        with:
-          channel-id: ${{ secrets.SLACK_CHANNEL_ID }}
-          payload: |
-            {
-              "text": "Deployment started (In Progress)",
-              "attachments": [
-                {
-                  "pretext": "Deployment started",
-                  "color": "dbab09",
-                  "fields": [
-                    {
-                      "title": "Status",
-                      "short": true,
-                      "value": "output value ${{ steps.check-update.outputs.errorOutput }}"
-                    }
-                  ]
-                }
-              ]
-            }
-        env:
-          SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
+      # - name: Slack Alert Notification
+      #   uses: slackapi/slack-github-action@v1.23.0
+      #   with:
+      #     channel-id: ${{ secrets.SLACK_CHANNEL_ID }}
+      #     payload: |
+      #       {
+      #         "text": "Deployment started (In Progress)",
+      #         "attachments": [
+      #           {
+      #             "color": "dbab09",
+      #             "fields": [
+      #               {
+      #                 "title": "Status",
+      #                 "short": true,
+      #                 "value": "output value ${{ steps.check-update.outputs.errorOutput }}"
+      #               }
+      #             ]
+      #           }
+      #         ]
+      #       }
+      #   env:
+      #     SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
