+++ package.json
@@ -5,13 +5,13 @@
   "main": "index.js",
   "scripts": {
     "dev": "ts-node src/index.ts cosmos/axelar-dojo.json",
+    "cron": "ts-node src/schedule.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
   "license": "",
   "dependencies": {
-    "@actions/core": "^1.10.0",
     "@keplr-wallet/background": "^0.12.0-alpha.0",
     "@keplr-wallet/chain-validator": "^0.12.0-alpha.0",
     "axios": "^0.27.2",
